Major update
I rewrote most parsing code, finding that the only use of the old code was to use as a navigation system through this mess. I decided to keep the mostly inefficient parsing system that collects strings and then parses them instead of doing it as it goes because I think it's just funny.