Organisation and compilation
In this update I have added more organised file systems, moving game playing away from Program.cs, removing Ids because of their redundancy, adding constants, etc. Along with this I've added compilation. This allows for full games generated with this program to be stored in a singular file, meaning that they can be shared easier and making the use of this program simpler.